# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kira Code API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/containers/info": {
      "get": {
        "tags": [
          "containers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ref",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse8"
                }
              }
            }
          }
        }
      }
    },
    "/containers/attempt-context": {
      "get": {
        "tags": [
          "containers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ref",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse9"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/handoff/init": {
      "post": {
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffInitPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse38"
                }
              }
            }
          }
        }
      }
    },
    "/auth/handoff/complete": {
      "get": {
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "handoff_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "app_code",
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "default": null
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "error",
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "default": null
            },
            "style": "form"
          }
        ]
      }
    },
...
    "/filesystem/directory": {
      "get": {
        "tags": [
          "filesystem"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse52"
                }
              }
            }
          }
        }
      }
    },
    "/filesystem/git-repos": {
      "get": {
        "tags": [
          "filesystem"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse53"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse8": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContainerInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContainerInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "ContainerInfo": {
        "type": "object",
        "properties": {
          "attempt_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "attempt_id"
        ]
      },
      "ApiResponse9": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspaceContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspaceContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "WorkspaceContext": {
        "type": "object",
        "properties": {
          "workspace": {
            "$ref": "#/components/schemas/Workspace"
          },
          "workspace_repos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RepoWithTargetBranch"
            }
          },
          "orchestrator_session_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        },
        "required": [
          "workspace",
          "workspace_repos"
        ]
      },
      "Workspace": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "task_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "container_ref": {
            "type": [
              "string",
              "null"
            ]
          },
          "branch": {
            "type": "string"
          },
          "setup_completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "archived": {
            "type": "boolean"
          },
          "pinned": {
            "type": "boolean"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "worktree_deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "branch",
          "created_at",
          "updated_at",
          "archived",
          "pinned",
          "worktree_deleted"
        ]
      },
      "RepoWithTargetBranch": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "path": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "setup_script": {
            "type": [
              "string",
              "null"
            ]
          },
          "cleanup_script": {
            "type": [
              "string",
              "null"
            ]
          },
          "archive_script": {
            "type": [
              "string",
              "null"
            ]
          },
          "copy_files": {
            "type": [
              "string",
              "null"
            ]
          },
          "parallel_setup_script": {
            "type": "boolean"
          },
          "dev_server_script": {
            "type": [
              "string",
              "null"
            ]
          },
          "default_target_branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "default_working_dir": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "target_branch": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "path",
          "name",
          "display_name",
          "parallel_setup_script",
          "created_at",
          "updated_at",
          "target_branch"
        ]
      },
...
      "HandoffInitPayload": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "return_to": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "return_to"
        ]
      },
      "ApiResponse38": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HandoffInitResponseBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HandoffInitResponseBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "HandoffInitResponseBody": {
        "type": "object",
        "properties": {
          "handoff_id": {
            "type": "string",
            "format": "uuid"
          },
          "authorize_url": {
            "type": "string"
          }
        },
        "required": [
          "handoff_id",
          "authorize_url"
        ]
      },
...
      "ApiResponse52": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectoryListResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectoryListResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "DirectoryListResponse": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DirectoryEntry"
            }
          },
          "current_path": {
            "type": "string"
          }
        },
        "required": [
          "entries",
          "current_path"
        ]
      },
      "DirectoryEntry": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "is_directory": {
            "type": "boolean"
          },
          "is_git_repo": {
            "type": "boolean"
          },
          "last_modified": {
            "type": [
              "integer",
              "null"
            ],
            "format": "uint64",
            "minimum": 0
          }
        },
        "required": [
          "name",
          "path",
          "is_directory",
          "is_git_repo"
        ]
      },
      "ApiResponse53": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/DirectoryEntry"
            }
          },
          "error_data": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/DirectoryEntry"
            }
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}